Add a worker that publishes X11 clipboard status
This is a re-implementation of the xclip clipboard worker from the
'config' branch.

The C clipboard worker is event-driven and doesn't actually request the
selection to check whether it's empty, instead considering the clipboard
populated if the current selection has an owner.  This is a lot more
efficient and more responsive than the polling/xclip approach.